+++ examples/simple-taro-fast-framework/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## 1.16.30
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - taro-fast-component-extra@3.0.20
+  - taro-fast-component-prism@3.0.21
+  - taro-fast-component@3.0.20
+  - taro-fast-framework@3.1.14
+  - taro-fast-common@3.1.14
+
 ## 1.16.29
 
 ### Patch Changes