+++ packages/taro-fast-component/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 3.0.20
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - taro-fast-common@3.1.14
+
 ## 3.0.19
 
 ### Patch Changes